RFC1 y 1

+++ src/main/java/uniandes/isis2304/parranderos/negocio/RFC1.java
@@ -1,30 +1,32 @@
 package uniandes.isis2304.parranderos.negocio;
 
+import java.math.BigDecimal;
+
 public class RFC1 {
 
-	private int idIps;
+	private BigDecimal idIps;
 	
-	private int numero;
+	private BigDecimal numero;
 
-	public RFC1(int idIps, int numero) {
+	public RFC1(BigDecimal idIps, BigDecimal numero) {
 		super();
 		this.idIps = idIps;
 		this.numero = numero;
 	}
 
-	public int getIdIps() {
+	public BigDecimal getIdIps() {
 		return idIps;
 	}
 
-	public void setIdIps(int idIps) {
+	public void setIdIps(BigDecimal idIps) {
 		this.idIps = idIps;
 	}
 
-	public int getNumero() {
+	public BigDecimal getNumero() {
 		return numero;
 	}
 
-	public void setNumero(int numero) {
+	public void setNumero(BigDecimal numero) {
 		this.numero = numero;
 	}
